% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beta}
\alias{beta}
\title{A subset of GSE199057 dataset for vignette demonstration}
\format{
A numeric matrix
}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi
}
\usage{
data(beta)
}
\value{
A numeric matrix
}
\description{
The GSE199057 Gene Expression Omnibus dataset contains 68 mucosa samples 
from non-colon-cancer patients, from which we randomly sampled 24. 
Methylation data were measured on EPIC arrays and after removal of 
sex chromosomes and SNPs loci, it contains 816 126 probes. 
Pre-processing can be found on the _methyLImp2_ simulation github page 
https://github.com/annaplaksienko/methyLImp2_simulation_studies.
Here we subset only a quarter of probes from two smallest chromosomes (18 and 21) 
for the sake of demonstration.
}
\keyword{datasets}
