% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beta_meta}
\alias{beta_meta}
\title{Metadata information for GSE199057 dataset for vignette demonstration}
\format{
A data.frame
}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi
}
\usage{
data(beta_meta)
}
\value{
A data.frame.
}
\description{
The GSE199057 Gene Expression Omnibus dataset contains 68 mucosa samples from non-colon-cancer patients, from which we randomly sampled 24. Methylation data were measured on EPIC arrays and after removal of sex chromosomes and SNPs loci, it contains 816 126 probes. We have subset only probes from two chromosomes (18 and 21) for the sake of demonstration.
}
\keyword{datasets}
