% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{custom_anno_example}
\alias{custom_anno_example}
\title{An example of how custom (user provided) annotation data frame should look like}
\format{
A data.frame
}
\usage{
data(custom_anno_example)
}
\value{
A data.frame.
}
\description{
A snippet from an annotation for 450K methylation dataset from ChAMPdata package. Only 5 CpGs are chosen simply to provide an example of the data frame organization.
}
\keyword{datasets}
