% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluatePerformance.R
\name{evaluatePerformance}
\alias{evaluatePerformance}
\title{Evaluate root mean square error (RMSE) and mean absolute error (MAE)}
\usage{
evaluatePerformance(beta_true, beta_imputed, na_positions)
}
\arguments{
\item{beta_true}{first numeric data matrix.}

\item{beta_imputed}{second numeric data matrix}
}
\value{
A number, root mean square error.
}
\description{
This function computes root mean square error and mean absolute error as an element-wise difference between two matrices (apart from NA elements): RMSE = sqrt(sum_i (beta_i - hat_beta_i)^2 / #artificial_NAs), MAE = sum_i |beta_i - hat_beta_i| / #artificial_NAs.
}
