% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluatePerformance.R
\name{extract_values}
\alias{extract_values}
\title{Extract a vector of beta values of interest out of a matrix, 
given the positions of artificial NAs}
\usage{
extract_values(beta, na_positions)
}
\arguments{
\item{beta}{numeric data matrix}

\item{na_positions}{a list where each element is a list of two elements: 
column id and ids of rows with NAs in that column (structure matches the 
output of generateMissingData function). We need this because some NAs in 
the dataset are from real data and not artificial, so we can't evaluate the 
performance of the method on them since we do not know real value. 
Therefore, we need to know the positions of artificial NAs.}
}
\value{
A numerical vector of beta values
}
\description{
Extract a vector of beta values of interest out of a matrix, 
given the positions of artificial NAs
}
\keyword{internal}
