% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{inv.plogit}
\alias{inv.plogit}
\title{Inverse of the pseudo-logit function.}
\usage{
inv.plogit(x, min = 0, max = 1)
}
\arguments{
\item{min}{a number, default value is 0}

\item{max}{a number, default value is 1}

\item{X}{a numeric matrix}
}
\description{
Inverse of the pseudo-logit function.
}
\keyword{internal}
