% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyLImp.R
\name{methyLImp}
\alias{methyLImp}
\title{Impute missing values in methylation dataset}
\usage{
methyLImp(
  data,
  type = c("450K", "EPIC", "user"),
  annotation = NULL,
  min = 0,
  max = 1,
  max.sv = NULL,
  col.list = NULL,
  ncores = NULL,
  minibatch_frac = 1,
  minibatch_reps = 1
)
}
\arguments{
\item{data}{a numeric data matrix with missing values, with samples in rows and variables (probes) in columns.}

\item{type}{a type of data. It is used to provide correct annotation: 450K, EPIC or user-provided. If the latter is chosen, annotation must be provided in the next argument.}

\item{annotation}{a data frame, user provided annotation. Must contain two columns: cpg and chr, and provide a match between CpG sites and chromosomes. Choose "user" in the previous argument to be able to provide user annotation.}

\item{min}{a number, minimum value for bounded-range variables. Default: 0 (we assume beta-value representation of the methylation data). Unrestricted range if \code{min} or \code{max = NULL}.}

\item{max}{a number, maximum value for bounded-range variables. Default: 1 (we assume beta-value representation of methylation data). Unrestricted range if \code{min} or \code{max = NULL}.}

\item{max.sv}{a number, maximum of a singular value to be used in the pseudoinverse. If \code{NULL}, use all singular values.}

\item{col.list}{a numeric vector, restricts the imputation only to the specified columns. If \code{NULL}, all columns are considered.}

\item{ncores}{number of cores to use in parallel computation. If \code{NULL}, set to \code{#physical cores} - 1.}

\item{minibatch_frac}{a number, what percentage of samples to use for mini-batch computation. The default is 1 (i.e., 100% of samples are used, no mini-batch).}

\item{minibatch_reps}{a number, how many times repeat computations with a fraction of samples (more times - better performance). The default is 1 (as a companion to default fraction of 100%. i.e. no mini-batch).}
}
\value{
A numeric matrix \code{out} with imputed data is returned.
}
\description{
This function performs missing value imputation specific for DNA methylation data. The method is based on linear regression since methylation levels show a high degree of inter-sample correlation. Implementation is parallelised over chromosomes since probes on different chromosomes are usually independent.
}
