% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyLImp2-package.R
\docType{package}
\name{methyLImp2-package}
\alias{methyLImp2-package}
\title{methyLImp2: Missing value estimation of DNA methylation data}
\description{
This package allows to estimate missing values in DNA methylation data. methyLImp method is based on linear regression since methylation levels show a high degree of inter-sample correlation. Implementation is parallelised over chromosomes since probes on different chromosomes are usually independent. Mini-batch approach to reduce the runtime in case of large number of samples is available.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/annaplaksienko/methyLImp2}
  \item Report bugs at \url{https://github.com/annaplaksienko/methyLImp2/issues}
}

}
\author{
\strong{Maintainer}: Anna Plaksienko \email{anna@plaxienko.com} (\href{https://orcid.org/0000-0001-9607-057X}{ORCID})

Authors:
\itemize{
  \item Pietro Di Lena (\href{https://orcid.org/0000-0002-1838-8918}{ORCID})
  \item Claudia Angelini (\href{https://orcid.org/0000-0001-8350-8464}{ORCID})
  \item Christine Nardini (\href{https://orcid.org/0000-0001-7601-321X}{ORCID})
}

}
\keyword{internal}
