% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyLImp_internal.R
\name{methyLImp_internal}
\alias{methyLImp_internal}
\title{Impute missing values in methylation dataset}
\usage{
methyLImp_internal(dat)
}
\arguments{
\item{dat}{a numeric data matrix with missing values, with samples in rows and variables (probes) in columns.}
}
\value{
A numeric matrix \code{out} with imputed data is returned.
}
\description{
This function performs missing value imputation specific for DNA methylation data. The method is based on linear regression since methylation levels show a high degree of inter-sample correlation.
}
\keyword{internal}
