% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_methyLImp2.R
\name{mod_methyLImp2_internal}
\alias{mod_methyLImp2_internal}
\title{Impute missing values in methylation dataset}
\usage{
mod_methyLImp2_internal(
  dat,
  min,
  max,
  skip_imputation_ids,
  minibatch_frac,
  minibatch_reps
)
}
\arguments{
\item{dat}{a numeric data matrix with missing values,
with samples in rows and variables (probes) in columns.}

\item{min}{a number, minimum value for bounded-range variables.
Default is 0 (we assume beta-value representation of the methylation data).
Can be user provided in case of other types of data.}

\item{max}{a number, maximum value for bounded-range variables.
Default is 1 (we assume beta-value representation of the methylation data).
Can be user provided in case of other types of data.}

\item{skip_imputation_ids}{a character vector of names of the columns (CpGs names)
for which \emph{not} to perform the imputation. If \code{NULL}, all columns
are considered.}

\item{minibatch_frac}{a number between 0 and 1, what fraction of samples
to use for mini-batch computation. Remember that if your data has several groups,
mini-batch will be applied to each group separately but with the same fraction,
so choose it accordingly. However, if your chosen fraction will be smaller
than a matrix dimension for some groups, mini-batch will be just ignored.
We advise to use mini-batch only if you have large number of samples,
order of hundreds. The default is 1 (i.e., 100\% of samples are used,
no mini-batch).}

\item{minibatch_reps}{a number, how many times to repeat computations with
a fraction of samples specified above (more times -> better performance but
more runtime). The default is 1 (as a companion to default fraction of 100\%,
i.e. no mini-batch).}
}
\value{
A numeric matrix \eqn{out} with imputed data is returned.
}
\description{
Impute missing values in methylation dataset
}
\keyword{internal}
