% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beta_meta}
\alias{beta_meta}
\title{Metadata information for GSE199057 dataset for vignette demonstration}
\format{
A data.frame
}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi
}
\usage{
data(beta_meta)
}
\value{
A data.frame.
}
\description{
The GSE199057 Gene Expression Omnibus dataset contains 68 mucosa samples 
from non-colon-cancer patients, from which we randomly sampled 24. 
Dataset contains metadata for those.
}
\keyword{datasets}
