% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{plogit}
\alias{plogit}
\title{Pseudo-logit function with range restricted to \eqn{[-X, X]} where \eqn{X} depends on the double machine precision.}
\usage{
plogit(x, min = 0, max = 1)
}
\arguments{
\item{x}{a numeric matrix}

\item{min}{a number, default value is 0}

\item{max}{a number, default value is 1}
}
\value{
A numeric matrix
}
\description{
Pseudo-logit function with range restricted to \eqn{[-X, X]} where \eqn{X} depends on the double machine precision.
}
\keyword{internal}
