% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_functions.R
\name{plogit}
\alias{plogit}
\title{Pseudo-logit function}
\usage{
plogit(X, min = 0, max = 1)
}
\arguments{
\item{X}{a numeric matrix}

\item{min}{a number, default value is 0}

\item{max}{a number, default value is 1}
}
\value{
A numeric matrix
}
\description{
Pseudo-logit function
}
\keyword{internal}
